+++ src/color.rs
@@ -92,7 +92,7 @@ impl Colors {
     }
 
     /// Predefined mapping of ANSI standard and extended colors.
-    const COLOR_MAPPINGS: [(&str, u8); 33] = [
+    const COLOR_MAPPINGS: [(&str, u8); 34] = [
         // --- ANSI colors ---
         ("black", 0),
         ("red", 1),
@@ -126,6 +126,7 @@ impl Colors {
         ("burnt-orange", 166),
         ("orchid-bloom", 175),
         ("amber-dusk", 180),
+        ("copper-flame", 208),
         ("peachy-orange", 216),
         ("golden-glow", 223),
     ];