+++ src/main.rs
@@ -57,12 +57,6 @@ use std::process::ExitCode;
 /// Usage documentation for display to terminal.
 const USAGE: &str = include_str!("include/usage.in");
 
-// Version and build information.
-const PACKAGE_NAME: &str = env!("CARGO_PKG_NAME");
-const PACKAGE_VERSION: &str = env!("CARGO_PKG_VERSION");
-const BUILD_HASH: &str = env!("BUILD_HASH");
-const BUILD_DATE: &str = env!("BUILD_DATE");
-
 /// Used for restoring the terminal via [`Drop`] to its original state.
 struct RestoreTerminal;
 
@@ -88,7 +82,10 @@ fn run() -> Result<()> {
         println!("{USAGE}");
         Ok(())
     } else if opts.version {
-        println!("{PACKAGE_NAME} {PACKAGE_VERSION} ({BUILD_HASH} {BUILD_DATE})");
+        println!("{}", etc::version());
+        Ok(())
+    } else if opts.source {
+        println!("{}", etc::SOURCE_URL);
         Ok(())
     } else if opts.keys {
         print!("{}", help::keys_content());